fix(ios): Show the build date in localtime This change also includes a small refactor to move the versioning convenience APIs onto a `UIApplication` extension.
